md2doc-service-python/pyproject.toml
[project]
name = "md2doc-mcp"
version = "1.0.0"
description = "Markdown to Word document converter MCP server"
authors = [
    {name = "md2doc-plus", email = "dev@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mcp>=0.1.0",
    "python-docx>=1.1.0",
    "Pillow>=10.0.0",
    "pyecharts>=2.0.0",
    "snapshot-pyppeteer>=0.0.2",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

